ppo/lab6/pyproject.toml
[tool.poetry]
name = "lab6"
version = "0.1.0"
description = ""
authors = ["Evgeniy Feder <dzaba67@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
